Add files via upload
Insurance_claim is a MYSQL project which encompasses the study of data of key players in the insurance industry that are customers, agents and vendors. To elaborate, the table customer draws all the important information regarding the customers, their policies, claims, incidents history and so on. The agents are the employees assigned with the task of managing claims, all the information regarding the agents is chalked up in the agents table. Lastly, the vendors are those enterprises or individuals who are given the responsibility to investigate into the claim whether they are legible or not, hence the vendor table contains the attributes that records the information of all the vendors.


The objective of this MYSQL project is to maintain the customer records, tracking the policies, managing claims, vendor management, risk assessment and fraud detection.